Refactor DTO namespaces and update references
Refactored DTO namespaces to align with a new folder structure, moving them from `Zucchinimvc.Models.DTOs` to service-specific namespaces under `Zucchinimvc.Application.Services.[ServiceName].DTOs`. Updated all affected files, including services, repositories, clients, and views, to use the new namespaces. Adjusted `Zucchinimvc.csproj` to reflect the updated file paths. Ensured consistency across Analytics, CMS, Currency, Searches, and Weather services.
