+++ ZucchiniMVC/Application/Services/Analytics/DTOs/AnalyticsSummaryDto.cs
@@ -1,4 +1,4 @@
-namespace Zucchinimvc.Models.DTOs.Analytic;
+namespace Zucchinimvc.Application.Services.Analytics.DTOs;
 
 public class AnalyticsSummaryDto
 {
