+++ ZucchiniMVC/Application/Services/Analytics/IAnalyticsService.cs
@@ -1,5 +1,5 @@
 using ZucchiniCore.enums;
-using Zucchinimvc.Models.DTOs.Analytic;
+using Zucchinimvc.Application.Services.Analytics.DTOs;
 
 namespace Zucchinimvc.Application.Services.Analytics;
