+++ ZucchiniMVC/Application/Services/Currency/DTOs/CurrencyRateResponse.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace Zucchinimvc.Models.DTOs.CurrencyDTOs
+namespace Zucchinimvc.Application.Services.Currency.DTOs
 {
     public class CurrencyRateResponse
     {
