+++ ZucchiniMVC/Infrastructure/ApiClients/AzureInsightClient/AzureInsightClient.cs
@@ -1,5 +1,5 @@
 using Microsoft.ApplicationInsights;
-using Zucchinimvc.Models.DTOs.Analytic;
+using Zucchinimvc.Application.Services.Analytics.DTOs;
 
 namespace Zucchinimvc.Infrastructure.ApiClients.AzureInsightClient
 {
