+++ ZucchiniMVC/Infrastructure/Repositories/SearchRepo/SearchRepository.cs
@@ -1,7 +1,7 @@
 ﻿
 using Azure.Search.Documents;
+using Zucchinimvc.Application.Services.Searches.DTOs;
 using Zucchinimvc.Infrastructure.ApiClients.ZucchininSearchClient;
-using Zucchinimvc.Models.DTOs.SearchDTOs;
 
 namespace Zucchinimvc.Infrastructure.Repositories.SearchRepo
 {
