+++ ZucchiniMVC/Views/Analytics/Index.cshtml
@@ -1,4 +1,4 @@
-@model Zucchinimvc.Models.DTOs.Analytic.AnalyticsSummaryDto
+@model Zucchinimvc.Application.Services.Analytics.DTOs.AnalyticsSummaryDto
 
 @{
     ViewData["Title"] = "Analytics Dashboard";
