+++ ZucchiniMVC/Views/Shared/Components/Search/Default.cshtml
@@ -1,4 +1,4 @@
-@model IEnumerable<Zucchinimvc.Models.DTOs.SearchDTOs.ArticlesSearchResultDTO>
+@model IEnumerable<Zucchinimvc.Application.Services.Searches.DTOs.ArticlesSearchResultDTO>
 
 <div class="dropdown-menu show w-100 shadow-lg" style="min-width: 100%; max-width: 500px; z-index: 1050; max-height: 450px; overflow-y: auto; border: 1px solid #e9ecef; border-radius: 0 0 8px 8px; border-top: none; top: 100% !important; left: 0 !important;">
     @if (Model.Any())
